Try to delete temp files sooner in the test to help builds from the command line on Windows.

+++ commons-vfs2/src/test/java/org/apache/commons/vfs2/impl/test/DefaultFileMonitorTest.java
@@ -57,7 +57,9 @@ public class DefaultFileMonitorTest extends AbstractVfsTestCase {
     @Override
     public void tearDown() throws Exception {
         if (testFile != null) {
-            testFile.deleteOnExit();
+            if (!testFile.delete()) {
+                testFile.deleteOnExit();
+            }
         }
         super.tearDown();
     }
